Fix StringIterator len() return type to int to match Go JNI interface signature

+++ android/app/src/main/java/com/shieldlink/vpn/MyVpnService.java
@@ -172,7 +172,7 @@ public class MyVpnService extends VpnService implements PlatformInterface, Comma
                 return null;
             }
             @Override
-            public long len() {
+            public int len() {
                 return 0;
             }
         };
